# [SQL storage indexer] Use bytes (blob) to store data in database instead of string