--- a/internal/database/model.go
+++ b/internal/database/model.go
@@ -17,6 +17,6 @@ type Package struct {
 	DiscoveryFields string
 	Type            string
 	Path            string
-	Data            string
-	BaseData        string
+	Data            []byte
+	BaseData        []byte
 }