This PR updates the type for `Data` and `BaseData` fields to be `[]byte` instead of `string`.
This change requires to use `BLOB` as a SQLite type.

As it is shown in the following benchmarks, it results in an improvement on `bytes/operation` and `allocations/operation`

- Get operation benchmark:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/package-registry/internal/storage
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                 │ ../package-registry-main/sql_prod_indexer_main_get.txt │ sql_prod_indexer_main_get_bytes.txt │
                 │                         sec/op                         │      sec/op        vs base          │
SQLIndexerGet-16                                               2.364 ± 8%         2.461 ± 15%  ~ (p=0.123 n=10)

                 │ ../package-registry-main/sql_prod_indexer_main_get.txt │ sql_prod_indexer_main_get_bytes.txt  │
                 │                          B/op                          │     B/op      vs base                │
SQLIndexerGet-16                                             476.8Mi ± 0%   418.9Mi ± 0%  -12.14% (p=0.000 n=10)

                 │ ../package-registry-main/sql_prod_indexer_main_get.txt │ sql_prod_indexer_main_get_bytes.txt │
                 │                       allocs/op                        │  allocs/op    vs base               │
SQLIndexerGet-16                                              6.020M ± 0%    5.945M ± 0%  -1.25% (p=0.000 n=10)
```
- Updating index operatoin benchmark:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/package-registry/internal/storage
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                         │ ../package-registry-main/sql_prod_indexer_main_update.txt │ sql_prod_indexer_main_update_bytes.txt │
                         │                          sec/op                           │        sec/op         vs base          │
SQLIndexerUpdateIndex-16                                                  19.80 ± 7%             19.49 ± 4%  ~ (p=0.143 n=10)

                         │ ../package-registry-main/sql_prod_indexer_main_update.txt │ sql_prod_indexer_main_update_bytes.txt │
                         │                           B/op                            │      B/op        vs base               │
SQLIndexerUpdateIndex-16                                                5.415Gi ± 0%      5.037Gi ± 0%  -6.99% (p=0.000 n=10)

                         │ ../package-registry-main/sql_prod_indexer_main_update.txt │ sql_prod_indexer_main_update_bytes.txt │
                         │                         allocs/op                         │    allocs/op     vs base               │
SQLIndexerUpdateIndex-16                                                 13.52M ± 0%       13.49M ± 0%  -0.16% (p=0.000 n=10)
```